requirements.txt
Django==5.0.6
djangorestframework==3.15.1
channels==4.0.0
celery==5.3.6
redis==5.0.6
openai==1.35.13
python-dotenv==1.0.1
daphne==4.1.2
channels_redis==4.1.0
httpx==0.27.0
# REMOVED: gevent==23.9.1
